requirements.txt
textx
fire
mermaid-py